added time to read calculation

+++ src/Services/IPostInfo.cs
@@ -2,6 +2,6 @@
 {
     public interface IPostInfo
     {
-        long GetTimeToRead(string postHtml);
+        double GetTimeToRead(string postHtml);
     }
 }
